refactor: move isolate theme & move ThemeProvider inst to index.js
