+++ src/theme.js
@@ -1,0 +1,19 @@
+import { createTheme } from "@material-ui/core/styles";
+import purple from "@material-ui/core/colors/purple";
+import green from "@material-ui/core/colors/green";
+
+const theme = createTheme({
+  palette: {
+    primary: {
+      main: purple[500],
+    },
+    secondary: {
+      main: green[500],
+    },
+    text: {
+      primary: green[500],
+    },
+  },
+});
+
+export default theme;
